Pictures/Janhavi/hawk-pro1/requirements.txt
fastapi
uvicorn[standard]
pydantic
numpy
pandas
scikit-learn
joblib
paho-mqtt
cryptography
pyjwt
web3
pytest
keras
tensorflow